MP P1 port (step 4d): fix CWinTcpNet abstract on POSIX — ConfigureSocket DWORD not u_long (LP64)
The override CWinTcpNet::ConfigureSocket(SOCKET,u_long) didn't match the base pure-virtual
CTcpNet::ConfigureSocket(SOCKET,DWORD) on LP64 (DWORD=uint32, u_long=64-bit; identical on
Win32) -> CWinTcpNet stayed abstract, 'new CWinTcpNet' failed. Now DWORD on both vp TCP +
IPX ConfigureSocket overrides. Resolves the abstract-class blocker; Windows behavior identical.
vdmplay.cpp 17->16; remaining = guard IPX/COMM transport-factory + the WSAAsyncSelect hidden
msg-window (CreateWindow/SetMessageQueue) #ifdef _WIN32, then step5 select() loop (the
WSAAsyncSelect->WM_WINSOCK->CWinTcpNet::OnMessage FD_* dispatch is the seam to drive from select).

+++ tools/vdmplay/vdmplay.cpp
@@ -252,8 +252,8 @@ public:
 
     void KeekNotifications( SOCKET s );
 
-    void ConfigureSocket( SOCKET s, u_long flags ) {
-        u_long on = 1;
+    void ConfigureSocket( SOCKET s, DWORD flags ) {   // DWORD not u_long: on LP64 they differ, so
+        u_long on = 1;                                 // u_long broke the override of the base pure-virtual (POSIX)
         int bufSize = 2048 * 4;
         int bvLen = sizeof( int );
 
@@ -460,8 +460,8 @@ public:
 
 
 
-    void ConfigureSocket( SOCKET s, u_long flags ) {
-        u_long on = 1;
+    void ConfigureSocket( SOCKET s, DWORD flags ) {   // DWORD not u_long: on LP64 they differ, so
+        u_long on = 1;                                 // u_long broke the override of the base pure-virtual (POSIX)
         int bufSize = 2048 * 4;
         int bvLen = sizeof( int );
